Removing Security Groups at AppScale Termination

+++ lib/common_functions.rb
@@ -798,6 +798,10 @@ module CommonFunctions
     CommonFunctions.get_from_yaml(keyname, :infrastructure)
   end
 
+  def self.get_group(keyname, required=true)
+    CommonFunctions.get_from_yaml(keyname, :group)
+  end
+
 
   def self.make_appscale_directory()
     # AppScale generates private keys for each cloud that machines are 
@@ -857,7 +861,7 @@ module CommonFunctions
     CommonFunctions.write_node_file(head_node_ip,
       head_node_result[:instance_id], options['table'],
       head_node_result[:secret_key], db_master_ip, all_ips,
-      options['infrastructure'], locations_yaml)
+      options['infrastructure'], options['group'], locations_yaml)
     remote_locations_file = "/root/.appscale/locations-#{keyname}.yaml"
     CommonFunctions.scp_file(locations_yaml, remote_locations_file,
       head_node_ip, head_node_result[:true_key])
@@ -865,13 +869,13 @@ module CommonFunctions
 
 
   def self.write_node_file(head_node_ip, instance_id, table, secret, db_master,
-    ips, infrastructure, locations_yaml)
+    ips, infrastructure, group, locations_yaml)
 
     infrastructure ||= "xen"
     tree = { :load_balancer => head_node_ip, :instance_id => instance_id , 
              :table => table, :shadow => head_node_ip, 
              :secret => secret , :db_master => db_master,
-             :ips => ips , :infrastructure => infrastructure }
+             :ips => ips , :infrastructure => infrastructure, :group => group }
     loc_path = File.expand_path(locations_yaml)
     File.open(loc_path, "w") {|file| YAML.dump(tree, file)}
   end
@@ -980,6 +984,11 @@ module CommonFunctions
       language = "java"
       # don't remove user's jar files, they may have their own jars in it
       #FileUtils.rm_rf("/tmp/#{temp_dir}/war/WEB-INF/lib/", :secure => true)
+      sdk_file = File.join(fullpath, "war/WEB-INF/lib/appengine-api-1.0-sdk-#{JAVA_AE_VERSION}.jar")
+      if !File.exists?(sdk_file)
+        raise AppEngineConfigException.new("Unsupported Java appengine version. Please " +
+                                               "recompile and repackage your app with Java appengine #{JAVA_AE_VERSION}.")
+      end
       FileUtils.mkdir_p("/tmp/#{temp_dir}/war/WEB-INF/lib")
       temp_dir2 = CommonFunctions.get_random_alphanumeric
       FileUtils.rm_rf("/tmp/#{temp_dir2}", :secure => true)
@@ -1491,7 +1500,7 @@ module CommonFunctions
   end
 
 
-  def self.terminate_via_infrastructure(infrastructure, keyname, shadow_ip, secret)
+  def self.terminate_via_infrastructure(infrastructure, keyname, group, shadow_ip, secret)
     Kernel.puts "About to terminate instances spawned via #{infrastructure} " +
       "with keyname '#{keyname}'..."
     Kernel.sleep(2)
@@ -1499,12 +1508,19 @@ module CommonFunctions
     acc = AppControllerClient.new(shadow_ip, secret)
     if acc.is_live?
       acc.kill()
-      # TODO(cgb): read the group from the locations.yaml file and delete that
-      cmd = "#{infrastructure}-delete-group appscale"
-      Kernel.puts CommonFunctions.shell("#{cmd}")
+      cmd = "#{infrastructure}-delete-group #{group} > /dev/null; echo $?"
+      while true
+        delete_group_output = CommonFunctions.shell("#{cmd}")
+        if delete_group_output.strip == "0"
+          break
+        else
+          Kernel.puts "Waiting for instances to shutdown"
+          Kernel.sleep(5)
+        end
+      end
       Kernel.puts "Terminated AppScale in cloud deployment."
     else
-      VMTools.terminate_infrastructure_machines(infrastructure, keyname)
+      VMTools.terminate_infrastructure_machines(infrastructure, keyname, group)
     end
   end
